+++ odoo/local-src/stock_delivery_note/models/stock_picking.py
@@ -218,7 +218,7 @@ class StockPicking(models.Model):
             ]
         )
 
-        vat_group = self.env.ref('stock_delivery_note.vat_tax_group')
+        vat_group = self.env.ref('specific_data.vat_tax_group')
         # The product lines
         grouped_lines = self.get_moves_by_order()
         for group in grouped_lines: